requirements.txt
PyQt5
reportlab
python-barcode
Pillow

